removed route files that were not being used
