+++ server/db/models/user.js
@@ -26,49 +26,11 @@ const User = db.define('user', {
   },
   profileUrl: {
     type: Sequelize.STRING
-  },
-  country: {
-    type: Sequelize.STRING
-  },
-  followers: {
-    type: Sequelize.INTEGER
-  },
-  product: {
-    type: Sequelize.TEXT
   }
 })
 
 module.exports = User
 
-// const User = db.define('user', {
-//   email: {
-//     type: Sequelize.STRING,
-//     unique: true,
-//     allowNull: false
-//   },
-//   password: {
-//     type: Sequelize.STRING,
-//     // Making `.password` act like a func hides it when serializing to JSON.
-//     // This is a hack to get around Sequelize's lack of a "private" option.
-//     get() {
-//       return () => this.getDataValue('password')
-//     }
-//   },
-//   salt: {
-//     type: Sequelize.STRING,
-//     // Making `.salt` act like a function hides it when serializing to JSON.
-//     // This is a hack to get around Sequelize's lack of a "private" option.
-//     get() {
-//       return () => this.getDataValue('salt')
-//     }
-//   },
-//   spotifyId: {
-//     type: Sequelize.STRING
-//   }
-// })
-/**
- * instanceMethods
- */
 User.prototype.correctPassword = function(candidatePwd) {
   return User.encryptPassword(candidatePwd, this.salt()) === this.password()
 }
